backend/requirements.txt
Django==3.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.6.0
Pillow==8.3.1
PyJWT==2.1.0



